+++ src/page/CheckList/components/CheckListItemRow.tsx
@@ -1,4 +1,4 @@
-import React, { ChangeEvent } from 'react';
+import React, { ChangeEvent, useMemo } from 'react';
 import { GrafanaTheme2 } from '@grafana/data';
 import { Checkbox, useStyles2 } from '@grafana/ui';
 import { css, cx } from '@emotion/css';
@@ -6,6 +6,7 @@ import { css, cx } from '@emotion/css';
 import { checkToUsageCalcValues, getCheckType } from 'utils';
 import { useUsageCalc } from 'hooks/useUsageCalc';
 import { AlertStatus } from 'components/AlertStatus/AlertStatus';
+import { getMissingCalNames, splitLabels } from 'page/CheckList/CheckList.utils';
 import { CheckItemActionButtons } from 'page/CheckList/components/CheckItemActionButtons';
 import { CheckListItemProps } from 'page/CheckList/components/CheckListItem';
 import { CheckListItemDetails } from 'page/CheckList/components/CheckListItemDetails';
@@ -15,6 +16,7 @@ import { DisableReasonHint } from 'page/CheckList/components/DisableReasonHint';
 export const CheckListItemRow = ({
   check,
   runtimeAlertState,
+  calNames,
   onLabelSelect,
   onTypeSelect,
   onStatusSelect,
@@ -24,6 +26,8 @@ export const CheckListItemRow = ({
   const styles = useStyles2(getStyles);
   const checkType = getCheckType(check.settings);
   const usage = useUsageCalc([checkToUsageCalcValues(check)]);
+  const { calLabels, customLabels } = useMemo(() => splitLabels(check.labels, calNames), [check.labels, calNames]);
+  const missingCalNames = useMemo(() => getMissingCalNames(check.labels, calNames), [check.labels, calNames]);
 
   return (
     <div
@@ -62,8 +66,9 @@ export const CheckListItemRow = ({
           activeSeries={usage?.activeSeries}
           probeLocations={check.probes.length}
           className={styles.listItemDetails}
-          labelCount={check.labels.length}
-          labels={check.labels}
+          labels={customLabels}
+          calLabels={calLabels}
+          missingCalNames={missingCalNames}
           onLabelClick={onLabelSelect}
           executionsRate={usage?.checksPerMonth}
         />